net: update False Start blacklist processing for \r\n
It's not clear what's happening because I don't have a Windows machine
and this was all done via try server runs:

I think the problem is that the new False Start blacklist is larger than
fread on Windows will read in one lump. That was causing the processing
to fail on Windows.

However, the simple fix of calling fread until its done didn't work
because it never reads enough bytes. What's happening is that ftell is
returning the true size of the file, but fread was converting \r\n to \n
so the reported number of bytes read was too small.

Why that wasn't happening before I'm not sure. I suspect that the file
on disk on the Windows builders is \n and it's patch on the tryserver
which is writing the file out with the extra \r bytes.

Either way, this hopefully makes everything work.

BUG=none
TEST=net_unittests

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@65421 0039d316-1c4b-4281-b951-d872f2087c98