Add support for --multi-install to mini_installer, fixed a memory leak, fixed use of Win32 registry API, and refactored SetFullInstallerFlag so that it can be exercised in unit tests (tests to follow).  Also explicitly turn off ExceptionHandling in the msvs_settings so that builds succeed in shared_library configs.
When --multi-install is not on command-line, "-full" is still added to the appropriate product "ap" value in official builds. When --multi-install is present, "-full" is added to the multi installer's "ap" value unless a product is being migrated from single-install to multi-install, in which case it is put on the product's "ap" value as before.

BUG=61609
TEST=Manual for now: run mini_installer.exe in various configurations, inspecting "ap" values as you go.
Review URL: http://codereview.chromium.org/6025003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@69710 0039d316-1c4b-4281-b951-d872f2087c98