Updating expectation of autofill sync tests when a null profile is added
In the past, it used to be possible to add an autofill profile that had its
label field populated with all other fields blank.

Today, if a sync test were to add an autofill profile with only a GUID
(leaving all other fields blank), the operation is expected to be unsuccessful.

This patch updates the sync test framework with this new expectation.

TBR=lipalani@chromium.org
BUG=69170
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/6230001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@70995 0039d316-1c4b-4281-b951-d872f2087c98